feat: Add getListStudentsIds function
This commit adds a new function `getListStudentsIds` to the `ES6_data_manipulation/1-get_list_student_ids.js` file. The function takes an array of student objects as input and returns an array of student IDs. This will be useful for retrieving a list of student IDs from an array of student objects.